实现将CSV中的参数量化为int8

+++ NNInference/quantization_util.py
@@ -13,7 +13,7 @@ class NNLayer(object):
         if (drtDataType == 'int8'):
             return max(abs(max_val), abs(min_val)) / (2**7 - 1)
         elif(drtDataType == 'int32'):
-            return min(abs(max_val), abs(min_val)) / (2**31 - 1)
+            return min(abs(max_val), abs(min_val)) / (2**31 - 1) # 原本 bais 使用，但会数据溢出
         else:
             raise ValueError(f"Unsupported data type: {drtDataType}. Valid types are 'int8' and 'int32'.")
 
@@ -33,7 +33,7 @@ class NNLayer(object):
         if self.weight is not None:
             self.weight = self.quantitate(self.weight, 'int8')
         if self.bias is not None:
-            self.bias = self.quantitate(self.bias, 'int32')
+            self.bias = self.quantitate(self.bias, 'int8') # 原本是 int32 ，但有溢出
 
 def show(dataArray):
     str_data = ",".join(map(str, dataArray))
